+++ test/_setup.test.js
@@ -4,15 +4,20 @@ testing = true;            // global
 VERBOSE = false;           // global
 
 global.Game = Object.create({});
-global.Creep = Object.create({});
 
-global.Creep.prototype = {};
+global.Creep = function() {
+  this.memory = {};
+  this.pos = {};
+  this.room = {};
+  this.carry = {};
+}
+
+Creep.prototype.pos = {};
 
 require('../src/proto/prototype.creep.js');
-global.Creep = Creep.prototype;
 
 describe('Mocha and Expect', () => {
   it('should properly run tests.', () => {
-    expect(1+3).toEqual(4);
+    expect(1 + 3).toEqual(4);
   });
 });