+++ test/timers.test.js
@@ -4,8 +4,10 @@ const timer = require('../src/timers').timer;
 
 describe('Timers', () => {
 
+  let  creep = new Creep();
+
   beforeEach(() => {
-    global.Creep.memory = {}
+    creep.memory = {}
   });
 
   describe('everyNTicks', () => {
@@ -20,61 +22,61 @@ describe('Timers', () => {
       expect(everyNTicks(17)).toEqual(true);
     });
 
-    it('should return true if Creep.memory.time not divible by n', () => {
-      Creep.ticksToLive = (5 * 4) + 1;
-      expect(everyNTicks.call(Creep, 5)).toEqual(false);
+    it('should return true if creep.memory.time not divible by n', () => {
+      creep.ticksToLive = (5 * 4) + 1;
+      expect(everyNTicks.call(creep, 5)).toEqual(false);
     });
 
-    it('should return true if Creep.memory.time divible by n', () => {
-      Creep.ticksToLive = (8 * 7);
-      expect(everyNTicks.call(Creep, 8)).toEqual(true);
+    it('should return true if creep.memory.time divible by n', () => {
+      creep.ticksToLive = (8 * 7);
+      expect(everyNTicks.call(creep, 8)).toEqual(true);
     });
   });
 
   describe('timer', () => {
     it('should start with n ticks', () => {
       const t = 10;
-      timer.call(Creep, t);
-      expect(Creep.memory.timer).toBe(t);
+      timer.call(creep, t);
+      expect(creep.memory.timer).toBe(t);
     });
 
     it('should count down tick by tick', () => {
       const t = 90;
-      timer.call(Creep, t);
-      for (let i = 0; i < 15; i++) timer.call(Creep, 10);
-      expect(Creep.memory.timer).toBe(t - 15);
+      timer.call(creep, t);
+      for (let i = 0; i < 15; i++) timer.call(creep, 10);
+      expect(creep.memory.timer).toBe(t - 15);
     });
 
     it('should return false if not done', () => {
       const t = 144;
-      timer.call(Creep, t);
-      expect(timer.call(Creep, t)).toBe(false);
+      timer.call(creep, t);
+      expect(timer.call(creep, t)).toBe(false);
     });
 
     it('should return true when done', () => {
       const t = 17;
-      timer.call(Creep, t);
-      for (let i = 0; i < t; i++) timer.call(Creep, t);
-      expect(timer.call(Creep, t)).toBe(true);
+      timer.call(creep, t);
+      for (let i = 0; i < t; i++) timer.call(creep, t);
+      expect(timer.call(creep, t)).toBe(true);
     });
 
     it('should restart the timer once donewhen repeat true', () => {
       const t = 17;
       const diff = 5;
       const repeat = true
-      timer.call(Creep, t, repeat);
-      for (let i = 0; i < t + diff; i++) timer.call(Creep, t, repeat);
-      expect(timer.call(Creep, t, repeat)).toBe(false);
-      expect(Creep.memory.timer).toBe(t - diff);
+      timer.call(creep, t, repeat);
+      for (let i = 0; i < t + diff; i++) timer.call(creep, t, repeat);
+      expect(timer.call(creep, t, repeat)).toBe(false);
+      expect(creep.memory.timer).toBe(t - diff);
     });
 
     it('should not restart the timer once done when repeat false', () => {
       const t = 17;
       const diff = 5;
-      timer.call(Creep, t);
-      for (let i = 0; i < t + diff; i++) timer.call(Creep, t);
-      expect(timer.call(Creep, t)).toBe(true);
-      expect(Creep.memory.timer).toBe(0)
+      timer.call(creep, t);
+      for (let i = 0; i < t + diff; i++) timer.call(creep, t);
+      expect(timer.call(creep, t)).toBe(true);
+      expect(creep.memory.timer).toBe(0)
     });
   });
 });